Move getScope, update logic, add tests for useUnit
