+++ packages/effector-react/index.ts
@@ -1,1 +1,2 @@
 export * from '../../src/react'
+export {Provider} from '../../src/react/scope'
