+++ packages/effector-react/scope.ts
@@ -1,1 +1,2 @@
 export * from '../../src/react/ssr'
+export {Provider} from '../../src/react/scope'
